+++ Tags/MandrillLib/CacheInterface/infinite/infinite_cache_file_reader.cpp
@@ -241,7 +241,7 @@ BCS_RESULT c_infinite_module_file_reader::associate_cache_cluster(c_infinite_cac
 	return BCS_S_OK;
 }
 
-BCS_RESULT c_infinite_module_file_reader::virtual_address_to_relative_offset(int64_t virtual_address, int32_t& relative_offset) const
+BCS_RESULT c_infinite_module_file_reader::virtual_address_to_relative_offset(long long virtual_address, long& relative_offset) const
 {
 	return BCS_E_UNSUPPORTED;
 	//BCS_RESULT rs = BCS_S_OK;
@@ -254,12 +254,12 @@ BCS_RESULT c_infinite_module_file_reader::virtual_address_to_relative_offset(int
 
 	//infinite::s_cache_file_tags_header tags_header = *reinterpret_cast<const infinite::s_cache_file_tags_header*>(buffer_info.begin);
 
-	//relative_offset = static_cast<int32_t>(virtual_address - tags_header.tag_instances_address + sizeof(infinite::s_cache_file_tags_header));
+	//relative_offset = static_cast<long>(virtual_address - tags_header.tag_instances_address + sizeof(infinite::s_cache_file_tags_header));
 
 	//return rs;
 }
 
-BCS_RESULT c_infinite_module_file_reader::page_offset_to_virtual_address(unsigned long page_offset, int64_t& virtual_address) const
+BCS_RESULT c_infinite_module_file_reader::page_offset_to_virtual_address(unsigned long page_offset, long long& virtual_address) const
 {
 	virtual_address = page_offset; // -cache_file_header->expected_base_address;
